user story 6: completed

+++ app/controllers/bookmarks_controller.rb
@@ -15,19 +15,37 @@ class BookmarksController < ApplicationController
   end
 
   def create
-    @bookmark = @topic.bookmarks.create( bookmark_params )
+    @bookmark = @topic.bookmarks.build( bookmark_params )
+    @bookmark.user = current_user
+
+    respond_to do |format|
+      if @bookmark.save
+        format.js
+      else
+        format.js { flash.now[:alert] = "There was an error creating the bookmark." }
+      end
+    end
   end
 
   def update
+    authorize @bookmark
+
     @bookmark.update( bookmark_params )
     redirect_to [@bookmark.topic, @bookmark]
   end
 
   def destroy
+    authorize @bookmark
+
     @topics = Topic.all.order('updated_at DESC').includes(:bookmarks)
 
-    if @bookmark.destroy
-      redirect_to @bookmark.topic
+    respond_to do |format|
+      if @bookmark.destroy
+        format.html { redirect_to @topic }
+        format.js
+      else
+        format.html { flash.now[:alert] = "There was an error deleting the bookmark." }
+      end
     end
   end
